Fix typo, link creation, creating new unnesessary redis
