+++ 3478/2/helpers/commands_helper.rb
@@ -1,0 +1,9 @@
+require 'require_all'
+require_rel '../utils/command/'
+
+module CommandsHelper
+  include StartCommand
+  include Unlink
+  include CheckoutCommand
+  include CheckinCommand
+end
