+++ 3478/2/utils/command/checkout_command.rb
@@ -6,6 +6,6 @@ module CheckoutCommand
   def checkout!(*)
     session[:check_type] = 'checkout'
     save_context :ask_for_photo
-    respond_with :message, text: '📸 Отправтье селфи'
+    respond_with :message, text: '📸 Отправте селфи'
   end
 end
